# Use custom struct instead of std::pair in QueryHeap