@SiarheiFedartsou 2024-05-30
> Nice readability improvement

I hope it is not only readability :) locally I have slight performance improvement on some of benchmarks. It can be confusing, but there are indeed some reasons why std::pair can be slower than raw struct: 
https://danlark.org/2020/04/13/why-is-stdpair-broken/
https://www.reddit.com/r/cpp/comments/ar4ghs/stdpair_disappointing_performance/

But tbh I am not 100% sure yet it is the case here.