@SiarheiFedartsou 2024-05-30
> e indeed some reasons why std::pair can be slower than raw struct: https://danlark.org/2020/04/13/why-is-stdpair-broken/ https://www.reddit.com/r/cpp/comments/ar4ghs/stdpair_disappointing_performance/
> 
> But tbh I am not 100% sure yet it is the case here.

Well, according to benchmark it seems to be indeed microoptimisation, but hard to say for sure as improvement is really small.